Update README.md
Wallets allow users to create transactions that sign ownership records of digital coins and assign them to new owners. The records of these exchanges make up the history of Bitcoin transactions, sometimes referred to as a ledger.



Transactions are recorded on a public ledger visible to all parties without the need for a financial institution’s involvement. This does not mean that banks will not use Bitcoin or be a part of the Bitcoin ecosystem. In fact, banks that use Bitcoin will have the advantages of its low cost transaction verification as a competing advantage in financial markets.